Corrected the comments.
